Add validation to updateRoles.
Do not add @Valid annotation to wine or review controllers because Wines and Reviews that are received by the controller are never valid. This is because Wine and Review entities must contain a User field, but I absolutely do not want to add an unnecessary user field in wine and review request bodies.

The downside of not checking for validity in controllers is that legitimately malformed request bodies get passed all the way to the DBMS before it rejects them and returns an error.

An obvious solution would be to use DTOs that are validated before they are mapped to models. I probably don't want to bother with creating separate DTOs, because the current system works fine as is.